Convert integration entry page to data table
